Fixes on Deputado table model, index() and initialize() methods are how fully kfunctional

+++ cidadao-de-olho/database/migrations/2019_06_21_015435_create_deputados_table.php
@@ -16,7 +16,7 @@ class CreateDeputadosTable extends Migration
         Schema::create('deputados', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->tinyInteger('id_almg')->nullable(false);
+            $table->integer('id_almg')->nullable(false);
             $table->string('name')->nullable(false);
             $table->string('partido')->nullable(false);
             $table->string('tag_localizacao')->nullable(false);
